+++ src/app.jsx
@@ -5,26 +5,7 @@ import CV from "./pages/cv"
 import Home from "./pages/home"
 import NotFound from "./pages/notFound"
 import Now from "./pages/now"
-import postGitHubTheme from "./posts/applying-github-dark-theme-miniflux-gpt.md"
-import postMonzo from "./posts/automating-simple-personal-finance-system-with-monzo.md"
-import postHiking from "./posts/hiking-bobotov-kuk-sedlo-zabljak-gpx.md"
-import postNowSummer2022 from "./posts/now/summer-autumn-2022.md"
-
-const posts = [
-  {
-    path: "/hiking-bobotov-kuk-sedlo-zabljak-gpx",
-    content: postHiking,
-  },
-  {
-    path: "/automating-simple-personal-finance-system-with-monzo",
-    content: postMonzo,
-  },
-  {
-    path: "/applying-github-dark-theme-miniflux-gpt",
-    content: postGitHubTheme,
-  },
-  { path: "/now/summer-autumn-2022", content: postNowSummer2022 },
-]
+import posts from "./posts"
 
 // TODO: Add a proper error component.
 const App = () => {
