+++ src/components/icon.jsx
@@ -1,0 +1,9 @@
+const Icon = () => {
+  return (
+    <div className="w-12 h-12 mb-4">
+      <img src="/icon.svg" alt="Icon" />
+    </div>
+  )
+}
+
+export default Icon
