+++ src/posts.jsx
@@ -1,0 +1,22 @@
+import postGitHubTheme from "./posts/applying-github-dark-theme-miniflux-gpt.md"
+import postMonzo from "./posts/automating-simple-personal-finance-system-with-monzo.md"
+import postHiking from "./posts/hiking-bobotov-kuk-sedlo-zabljak-gpx.md"
+import postNowSummer2022 from "./posts/now/summer-autumn-2022.md"
+
+const posts = [
+  {
+    path: "/hiking-bobotov-kuk-sedlo-zabljak-gpx",
+    content: postHiking,
+  },
+  {
+    path: "/automating-simple-personal-finance-system-with-monzo",
+    content: postMonzo,
+  },
+  {
+    path: "/applying-github-dark-theme-miniflux-gpt",
+    content: postGitHubTheme,
+  },
+  { path: "/now/summer-autumn-2022", content: postNowSummer2022 },
+]
+
+export default posts
